chore(faceted-search): Rename usage tracking tags
