+++ packages/faceted-search/src/components/FacetedToolbar/FacetedToolbar.component.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Toggle from '@talend/react-components/lib/Toggle';
 import { getTheme } from '@talend/react-components/lib/theme';
 
-import { FACETED_MODE, PENDO_TAGS } from '../../constants';
+import { FACETED_MODE, USAGE_TRACKING_TAGS } from '../../constants';
 import cssModule from './FacetedToolbar.scss';
 
 const theme = getTheme(cssModule);
@@ -13,12 +13,12 @@ const SwitchFacetedMode = ({ facetedMode, id, onChange, t }) => {
 		{
 			value: FACETED_MODE.BASIC,
 			label: t('FACETED_SEARCH_BASIC', { defaultValue: 'Basic' }),
-			dataFeature: PENDO_TAGS.BASIC,
+			dataFeature: USAGE_TRACKING_TAGS.BASIC,
 		},
 		{
 			value: FACETED_MODE.ADVANCED,
 			label: t('FACETED_SEARCH_ADVANCED', { defaultValue: 'Advanced' }),
-			dataFeature: PENDO_TAGS.ADVANCED,
+			dataFeature: USAGE_TRACKING_TAGS.ADVANCED,
 		},
 	];
